source/todolist-back/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0.3
djangorestframework==3.11.0
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.1
flake8==3.7.9
importlib-metadata==1.5.0
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-randomly==3.2.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
wcwidth==0.1.8
zipp==3.0.0
